view cart button added

+++ carts/views.py
@@ -77,13 +77,28 @@ def remove_cart(request, product_id):
         cart_item.delete()
     return redirect("cart")
 
-def remove_cart_item(request,product_id):
-    cart = Cart.objects.get(cart_id =_cart_id(request))
-    product = get_object_or_404(Product, id = product_id)
-    cart_item = CartItem.objects.get(product=product, cart=cart)
-    cart_item.delete()
-    return redirect("cart")
+def remove_cart(request, product_id, cart_item_id):
+    cart = Cart.objects.get(cart_id=_cart_id(request))
+    product = get_object_or_404(Product, id=product_id)
+    try:
+        # Use the unique cart_item_id instead of just the product
+        cart_item = CartItem.objects.get(product=product, cart=cart, id=cart_item_id)
+        if cart_item.quantity > 1:
+            cart_item.quantity -= 1
+            cart_item.save()
+        else:
+            cart_item.delete()
+    except:
+        pass
+    return redirect('cart')
 
+def remove_cart_item(request, product_id, cart_item_id):
+    cart = Cart.objects.get(cart_id=_cart_id(request))
+    product = get_object_or_404(Product, id=product_id)
+    # Use the unique cart_item_id here as well
+    cart_item = CartItem.objects.get(product=product, cart=cart, id=cart_item_id)
+    cart_item.delete()
+    return redirect('cart')
 
 def cart(request, total=0,quantity=0,cart_items=None):
     try:
